Improve Sagasu design & indexing
Now the hashes are collected in a central database (location to be
improved) similar to FFXIV Explorer. This database needs to be generated
once and doesn't have to be regen every time Sagasu is opened like
before. This indexer currently is a separate program.

Also adds a feature to extract files from the file tree window.