Fix searching in the Data Explorer
I don't know when/how this broke, but it was driving me crazy. Searching
for a specific file would break the model, causing it to "loop back" to
showing the root folders in children. I just had to make sure the row
property was set correctly in the TreeInformation structs.