Remove the oft unused Game Launcher and Argcracker
These are the least used out of all of my tools, and don't really belong
in this suite. Argcracker is so simple it can be shipped separately (and
doesn't really need to be C++.) Game Launcher is better handled by your
launcher, and the one I use - Astra - already has a feature to quickly
launch the game.