Even more index file improvements for Data Explorer
Now the index2 files are read in conjunction with the normal ones, this
is normally useless but some versions of the game (like the benchmark)
only ship these. We will attempt to de-duplicate any entries though, and
although it increases start-up time I think its negligible.

I also tried squeezing out a bit more speed in our start-up process so
it should be on-par as before or a bit shorter. I forgot to set an
interval on the search timer too, so that's also fixed.