Add a PathEdit control to quickly explore game files
This is similar to the existing line edits we used for game paths
already, but with a button to open Data Explorer. You can use this to
extract the file if needed, view its raw hex and so on.

Right now I use it for the model path in Armoury, but I'll replace other
instances of game paths with this control.